+++ sdk/containerservice/arm-containerservice/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
     
-## 16.1.0-beta.2 (2022-05-18)
+## 16.1.0-beta.3 (2022-05-19)
     
 **Features**
 
@@ -10,6 +10,7 @@
   - Added operation ManagedClusters.beginRotateServiceAccountSigningKeys
   - Added operation ManagedClusters.beginRotateServiceAccountSigningKeysAndWait
   - Added Interface AzureKeyVaultKms
+  - Added Interface ManagedClusterAutoReconcileProfile
   - Added Interface ManagedClusterIngressProfile
   - Added Interface ManagedClusterIngressProfileWebAppRouting
   - Added Interface ManagedClusterOidcIssuerProfile
@@ -25,6 +26,7 @@
   - Added Interface ManagedClusterSnapshotsListOptionalParams
   - Added Interface ManagedClusterSnapshotsUpdateTagsOptionalParams
   - Added Interface ManagedClustersRotateServiceAccountSigningKeysOptionalParams
+  - Added Interface ManagedClusterStorageProfileBlobCSIDriver
   - Added Interface NetworkProfileForSnapshot
   - Added Interface TrustedAccessRole
   - Added Interface TrustedAccessRoleBindingListResult
@@ -37,6 +39,7 @@
   - Added Interface TrustedAccessRoleRule
   - Added Interface TrustedAccessRolesListNextOptionalParams
   - Added Interface TrustedAccessRolesListOptionalParams
+  - Added Type Alias KeyVaultNetworkAccessTypes
   - Added Type Alias ManagedClusterSnapshot
   - Added Type Alias ManagedClusterSnapshotsCreateOrUpdateResponse
   - Added Type Alias ManagedClusterSnapshotsGetResponse
@@ -67,6 +70,7 @@
   - Interface ManagedClustersDeleteOptionalParams has a new optional parameter ignorePodDisruptionBudget
   - Interface ManagedClusterSecurityProfile has a new optional parameter azureKeyVaultKms
   - Interface ManagedClusterSecurityProfile has a new optional parameter workloadIdentity
+  - Interface ManagedClusterStorageProfile has a new optional parameter blobCSIDriver
   - Interface ManagedClusterStorageProfileDiskCSIDriver has a new optional parameter version
   - Class ContainerServiceClient has a new parameter managedClusterSnapshots
   - Class ContainerServiceClient has a new parameter trustedAccessRoleBindings
@@ -82,6 +86,8 @@
   - Type Alias ManagedCluster has a new parameter enableNamespaceResources
   - Type Alias ManagedCluster has a new parameter storageProfile
   - Type Alias ManagedCluster has a new parameter ingressProfile
+  - Type Alias ManagedCluster has a new parameter autoReconcileProfile
+  - Added Enum KnownKeyVaultNetworkAccessTypes
   - Added Enum KnownNetworkPluginMode
   - Added Enum KnownTrustedAccessRoleBindingProvisioningState
   - Enum KnownNetworkPlugin has a new value None